prevencao de deletes sem utilizador logado e adicao de include nos servicos de submissao de informacao

+++ core/core_action.php
@@ -6,38 +6,55 @@ $action = (isset($_GET["action"]) ? $_GET["action"] : $_POST["action"]);
 
 switch($action){
 	case "delete" : {
-		require_once "../data/" . $_GET["class"] ."_data.php";
 		
-		//chamada dinamica da funcao
-		$instance = new $_GET["class"];
-		$toReturn = $instance->delete($_GET["class"] . "_id = '" . $_GET["id"] . "'");
-		
-		if(isset($_GET["class"])){
+		require_once "../data/user_data.php";
 
-			$parameters = '';
-			$parametersStartIndex = strpos($_SERVER['HTTP_REFERER'], '?');
-			if ($parametersStartIndex !== false) {
-				$parameters = substr($_SERVER['HTTP_REFERER'], $parametersStartIndex + 1);
-				$parameters = str_replace("success=1", "", $parameters);
-				$parameters = str_replace("success=2", "", $parameters);
+		$newLocation = '';
+		$parameters = '';
+		$userData = new User();
 
-				if (substr($parameters, -1) === '&') {
-					$parameters =substr($parameters, 0,-1);
-				}
+		$parametersStartIndex = strpos($_SERVER['HTTP_REFERER'], '?');
+		if ($parametersStartIndex !== false) {
+			$parameters = substr($_SERVER['HTTP_REFERER'], $parametersStartIndex + 1);
+			$parameters = str_replace("success=1", "", $parameters);
+			$parameters = str_replace("success=2", "", $parameters);
 
-				if (substr($parameters, 0, 1) === '&') {
-					$parameters =substr($parameters, 1);
-				}
+			if (substr($parameters, -1) === '&') {
+				$parameters = substr($parameters, 0,-1);
 			}
-	
-			if (isset($_GET["redirect"])) {
-				header('Location: ' . $_GET["redirect"] . '?' . ($parameters != '&' ? $parameters : '') . '&success=' . ($toReturn == 1 ? '2' : $toReturn));
-			} else {
-				header('Location: /lists/' . $_GET["class"] . '-list.php?' . ($parameters != '&' ? $parameters : '') . '&success=' . ($toReturn == 1 ? '2' : $toReturn));
+
+			if (substr($parameters, 0, 1) === '&') {
+				$parameters = substr($parameters, 1);
 			}
-		} else {
-			header('Location: index.php');
 		}
+		
+		if ($userData->validateActiveUser()) {
+
+			require_once "../data/" . $_GET["class"] ."_data.php";
+
+			//chamada dinamica da funcao
+			$instance = new $_GET["class"];
+			$toReturn = $instance->delete($_GET["class"] . "_id = '" . $_GET["id"] . "'");
+
+			if(isset($_GET["class"])){
+
+				if (isset($_GET["redirect"])) {
+					$newLocation = $_GET["redirect"] . '?' . ($parameters != '' && $parameters != '&' ? $parameters . '&' : '') . 'success=' . ($toReturn == 1 ? '2' : $toReturn);
+				} else {
+					$newLocation = '/lists/' . $_GET["class"] . '-list.php?' . ($parameters != '' && $parameters != '&' ? $parameters . '&' : '') . 'success=' . ($toReturn == 1 ? '2' : $toReturn);
+				}
+			} else {
+				$newLocation = 'index.php';
+			}
+
+		} else {
+			$newLocation = '/lists/' . $_GET["class"] . '-list.php?' . ($parameters != '' && $parameters != '&' ? $parameters . '&' : '') . 'success=-1&reason=Não existe nenhum utilizador com login activo.';
+		}
+
+
+
+		header('Location: ' . $newLocation);
+		
 		break;
 	}
